Implemented DB methods for cancel and create

+++ src/main/java/dk/lsd/datalayer/DatabaseImpl.java
@@ -83,15 +83,15 @@ public class DatabaseImpl {
         }
     }
 
-    public void createRoomBooking(long dateFrom, long dateTo, int roomId, int bookingId) throws SQLException {
+    public void createRoomBooking(long dateFrom, long dateTo, long roomId, long bookingId) throws SQLException {
         var sql = "insert into room_booking(date_of_arrival, date_of_departure, room_id, booking_id) values (?,?,?,?)";
         try (var con = getConnection();
             var stmt = con.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
 
             stmt.setFloat(1, dateFrom);
             stmt.setFloat(2, dateTo);
-            stmt.setInt(3, roomId);
-            stmt.setInt(4, bookingId);
+            stmt.setLong(3, roomId);
+            stmt.setLong(4, bookingId);
 
             stmt.executeUpdate();
